requirements.txt
Django
Jinja2
requests
six
suds_jurko
